credit_judge_v2/requirements.txt
# Dependencies for the Credit Judge v2 application

# For making API calls to OpenAI
openai

# For loading environment variables from a .env file
python-dotenv

# For NLP analysis
textblob
textstat
